praas-app: wip commit post css decisions...
There are endless resources and discussions on what
is the best way to incorporate CSS flow into a react
based webapp. Rather than get religious and take sides,
we sat down and discussed what works for us. And then
made a decision that we think is objective, contextual
and facilitates more parallelism.

CSS design falls under visual domain. While we are
comfortable and good at writing it ourselves, that
is not our core strength as software developers.

Secondly, we are building an application for which
we are able to control the whole visual design. Put
differently this project is not about producing widgets
for others to consume in their application.

Taking the above two points into consideration we wanted
an approach that allows us to simply focus on code
and leave the visual aspects to those who can do it
better than us. Further when such a design doesn't
completely match our needs in certain areas, we wanted
the ability to tweek it without breaking the rest of
the application, using modular css.

Lastly we wanted to adopt a design system that is minimal,
documented and easily dissectable if required.

In this commit, we are bringing in mini.css with a few
modifications to support our webpack based build setup.

NOTE:
we can still build custom styled modular widgets that
don't bleed or break the rest of the application using
the primitives provided by mini.css and using sass
includes and extends.

TODO:
- figure out how to get css optimizations to work again
- make sure when all is said and done, the final css
  file for the app matches the size prior to this
  commit (~6KiB). At the moment it is much bigger...

Onwards...