Remove snapshot-it and test the right way
In hindsight snapshot based testing is no testing at all.
It provides very minimal confidence for catching regression
since it invariably desensitizes the developer from paying
attention to what has changed due to its wide scope.

The proper way to test is to write expect statements on values
that are important and leave out those that are don't care.