praas-app: remove lock files until we reach alpha
We had a policy to not commit lock files but then it looks like I accidentally
committed it around June/2019. This fix deletes package-lock.json

Aside: the reason we don't commit lock files in the early stages of a project
is because initially npm had issues where it would not allow package.json to
override package-lock.json leading to wasted time of developers in trying to
figure out what's going wrong. This has been fixed in npm > 5.1 but still
not worth committing because we cannot be sure of the version in use by
a developer checking out this project, especially since it is open source.

Read more at https://medium.com/coinmonks/everything-you-wanted-to-know-about-package-lock-json-b81911aa8ab8