proxy-server, test: caught an unhandled `TypeError`
when a PUT / PATCH / POST request is made without a body, the throws a
`TypeError: Cannot read property '0' of undefined` and dies.
This is because the application is not validating whether the request
has a body or not but just goes about processing the body of the
request.

This changeset fixes the error as well as implements the test for it.

Signed-off-by: shine <shine@m5nv.com>