proxy-server, test: PATCH existing entry in AirTable
the reason why the request is not generalized like the other requests
( at the beginning of the test suite ) is because the request requires
the ID ( `recordId` ) of an existing entry from the table which is to be
patched. This arbitrary record is inserted only when the AirTable test
context is entered, hence requiring the request to be within that scope.

tracking GitHub issue #120

Signed-off-by: shine <shine@m5nv.com>