proxy-server, chore: optimize hiddenFormField.include test
there was an `if` condition that was only checking for the length of the
records in the request and response. that was unnecessary and could be
optimized by using the `expect` directive from ChaiJS ( as seen in `Testing
Airtable Gateway should DELETE multiple entries` ). this is a test after
all.

Signed-off-by: shine <shine@m5nv.com>