crud-server, fix: Sequelize.findById in `PUT /user`
the `findById` alias was deprecated in Sequelize v5. the recommended
alternative is `findByPk`.
reference : https://sequelize.org/v5/manual/upgrade-to-v5.html#Model
this was missed during the upgrade of sequelize from v4 to v5 in a14cd84

also added extra `expect`ations for `PUT /user` test when the test fails.

Signed-off-by: shine <shine@m5nv.com>