crud-server, feat: create a test user via dotEnv file
the tests were running with dummy users that were generated dynamically
during the invocation of the tests. while this is ok in the scope of the
tests alone, it makes more sense to use a user that is specified by the
tester via a `.env` file. this also enables easy testing of the front-end
application as well since this avoids the 2 pre-requisite steps for
front-end testing - create a user ( after each database flush ), and
create multiple conduits. this manual process is also time consuming,
hence warranting for an automated process to set this up so that the
tester can get started sooner.

move creation of test conduits for the proxy server from `routes.test.js`
to `model.test.js` since there are more conduits being generated for the
user and also because it is easier to write directly into the database
than make a REST API request for the creation of multiple conduits.

changes ( bullets ) :
- move `dotEnv` configuration to `model.test.js`
- rename `.env.conduit` to `.env.conduit-user`
- move creation of test conduits for proxy server to `model.test.js`
- create new user from dotEnv
- use the newly created user for the test suite
- use a base object for common conduit parameters for proxy server
- add 200 dummy conduits to user for easier front-end testing

tracking GitHub issue #140

Signed-off-by: shine <shine@m5nv.com>