crud-server, refactor: normalize error messages and error shape; 1/n
The shape of the error in early codebase used the field or
object name as the key and the fault as the value.

Newer code, especially unhappy path tests were written to match
the shape of the Sequelize*Error object without normalizing.

Since the majority of codebase uses {field|entity-name : message }
format, the integration tests were modified to conform to the
predominant style. Sequelize errors are remapped to match.

Make error messages simpler and consistent along the way.