proxy-server, docs: add instructions for Airtable
The test suite has some hard-coded dependencies that rely on the Airtable
Base being used for the testing of the Airtable functionality. This is
because Airtable currently does not allow the creation, updation or
deletion of Airtable Bases themselves. They are only allowed on records
that go into an existing Airtable Base. So, the Airtable Base with the
the minimum required fields should be created by the tester themselves
to get the Airtable tests to run without failures.

Signed-off-by: shine <shine@m5nv.com>