crud-server, test: refactor tests by code functionality
the current state of the test suite of the POST method API looks a lot
more like a model test than an API test. this is because the tests checks
every field in the request ( multiple times ) extensively and exhaustively

this changeset refactors the tests to be smart and cover as many lines of
code testing all the functionality without the need for an extensive and
exhaustive test suite.

this refactor was driven by code coverage. focus was put into covering
as many lines of code in a single test eliminating the need for more
tests that would make redundant calls to the same lines of code. test
code coverage has been maintained to be exactly the same before and after
the refactor.

tracking GitHub issue #158

Signed-off-by: shine <shine@m5nv.com>