crud-server, refactor: remove un-unecessary function
to fetch the user object from the response object.
the abstraction function made it difficult to read the `expect` functions
in the test suite. `expect(res.body.user.email)` makes it more obvious
to the read and understand the expectation of the test than
`expect(User(res).email)` which would require the reader to look at the
definition for the function `User(res)`.
also, a mix of both `expect(res.body)` and `expect(User(res))` in the
test would not look good in terms of consistency in code style.

Signed-off-by: shine <shine@m5nv.com>