crud-server: add validation for array items
resource access control methods and hidden form field validation errors
generated by Sequelize produced an array of errors if multiple
validation criteria were violated. the tests only validated the first
element of the array ( or in the case of hidden form field, none ).

while the ideal solution would be https://github.com/chaijs/chai-things,
we probably do not want to add another package to the dependency list
when it is possible with a very simple ( not so complicated ) workaround

this changeset validates all elements of the array correctly using
`Array.prototype.forEach` as suggested in https://github.com/chaijs/chai/issues/410#issuecomment-348751229.

Signed-off-by: shine <shine@m5nv.com>