conduits.app: show some love 16b/n
summary:
  - signup page now ported to react-hook-form

NOTES:
1. react-hook-form's yup resolver behaves differently from Formik.
The order of validations is the order in which react-hook-form
fires. So if a field is required, user tabs out of that field,
then no error message will be presented if there is a previous
rule prior to the 'required' rule, e.g. min-length.

Formik is 'smarter' in figuring out that a 'required' validation
should fire first onBlur before checking for content. When time
permits should file a bug report with react-hook-form?

2. Talking about bugs, the other weirdness with react-hook-form
and react-testing-library combo is that we should be careful
in caching DOM elements. Some kind of proxying is happening
underneath these libraries that can result in non-deterministic
behavior with cached elements. See SO link below that other
developers have had issues akin to what I experienced.

A write up of all the issues encountered so far can be found at:
https://github.com/react-hook-form/react-hook-form/discussions/2620#discussioncomment-53028

references:
- https://stackoverflow.com/questions/62157528/waitforelementtoberemoved-is-timing-out-when-passed-an-element